Re-factor mdspan support for the view engine partial specializations
